// src/data/generated/smallbodies.ts
export interface SmallBodyData {
  name: string
  /** Absolute magnitude -- the size proxy used for the render radius. */
  h: number
  /** Semi-major axis, AU. */
  a: number
  e: number
  /** Inclination to the ecliptic, degrees. */
  inc: number
  /** Longitude of ascending node, degrees. */
  node: number
  /** Argument of perihelion, degrees. */
  argPeri: number
  /** Mean anomaly at epoch, degrees. */
  m0: number
  epoch: number
  group: string
}